[refactor] 경로 이동
