+++ src/main/java/calculator/domain/Number.java
@@ -1,4 +1,4 @@
-package domain;
+package calculator.domain;
 
 import java.util.Objects;
