+++ src/main/java/calculator/domain/Operator.java
@@ -1,4 +1,4 @@
-package domain;
+package calculator.domain;
 
 public interface Operator {
     Number operate(Number first, Number second);
