+++ src/main/java/calculator/domain/OperatorFactory.java
@@ -1,4 +1,4 @@
-package domain;
+package calculator.domain;
 
 import java.util.HashMap;
 import java.util.Map;
